+++ WireAnalytics/Sources/WireAnalytics/AnalyticEvents/Segmentation/AnalyticsValue.swift
@@ -16,9 +16,44 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-/// A protocol that defines a way to represent a value for analytics purposes.
+import Foundation
+
+/// A value that can be tracked.
+
 public protocol AnalyticsValue {
 
+    /// A representation suitable for tracking.
+
     var analyticsValue: String { get }
+
+}
+
+extension Bool: AnalyticsValue {
+
+    public var analyticsValue: String {
+        self ? "True" : "False"
+    }
+
+}
+
+
+extension UInt: AnalyticsValue {
+
+    public var analyticsValue: String {
+        String(logRound())
+    }
+
+    /// Rounds the integer value logarithmically to protect the privacy of BI data.
+    ///
+    /// The `logRound` method rounds numeric values into buckets of increasing size.
+    /// This logarithmic rounding means that smaller numbers are only slightly rounded,
+    /// whereas larger numbers are rounded more significantly. This approach helps to
+    /// protect privacy by reducing the precision of the values in a controlled manner.
+    ///
+    /// - Returns: A rounded integer value based on the base-2 logarithm of the original value.
+
+    func logRound() -> UInt {
+        UInt(log2(Double(self)).rounded())
+    }
 
 }
